Module boa_gc::cell

A garbage collected cell implementation

Structs

BorrowError

An error returned by GcCell::try_borrow.

BorrowFlag 🔒

BorrowFlag represent the internal state of a GcCell and keeps track of the amount of current borrows.

BorrowMutError

An error returned by GcCell::try_borrow_mut.

GcCell

A mutable memory location with dynamically checked borrow rules that can be used inside of a garbage-collected pointer.

GcCellRef

A wrapper type for an immutably borrowed value from a GcCell<T>.

GcCellRefMut

A wrapper type for a mutably borrowed value from a GcCell<T>.

Enums

BorrowState 🔒

BorrowState represents the various states of a BorrowFlag

Constants

BORROWFLAG_INIT 🔒

The base borrowflag init is rooted, and has no outstanding borrows.

ROOT 🔒

UNUSED 🔒

WRITING 🔒