Struct boa_gc::cell::GcRef

pub struct GcRef<'a, T: ?Sized + 'static> {
    flags: &'a Cell<BorrowFlag>,
    value: &'a T,
}

A wrapper type for an immutably borrowed value from a GcCell<T>.

Fields
flags: &'a Cell<BorrowFlag>value: &'a T
Implementations

impl<'a, T: ?Sized> GcRef<'a, T>

pub fn clone(orig: &GcRef<'a, T>) -> GcRef<'a, T>

Copies a GcCellRef.

The GcCell is already immutably borrowed, so this cannot fail.

This is an associated function that needs to be used as GcCellRef::clone(...). A Clone implementation or a method would interfere with the use of c.borrow().clone() to clone the contents of a GcCell.

pub fn try_map<U, F>(orig: Self, f: F) -> Option<GcRef<'a, U>>where U: ?Sized, F: FnOnce(&T) -> Option<&U>,

Tries to make a new GcCellRef from a component of the borrowed data, returning None if the mapping function returns None.

The GcCell is already immutably borrowed, so this cannot fail.

This is an associated function that needs to be used as GcCellRef::try_map(...). A method would interfere with methods of the same name on the contents of a GcCellRef used through Deref.

pub fn map<U, F>(orig: Self, f: F) -> GcRef<'a, U>where U: ?Sized, F: FnOnce(&T) -> &U,

Makes a new GcCellRef from a component of the borrowed data.

The GcCell is already immutably borrowed, so this cannot fail.

This is an associated function that needs to be used as GcCellRef::map(...). A method would interfere with methods of the same name on the contents of a GcCellRef used through Deref.

pub fn map_split<U, V, F>(orig: Self, f: F) -> (GcRef<'a, U>, GcRef<'a, V>)where U: ?Sized, V: ?Sized, F: FnOnce(&T) -> (&U, &V),

Splits a GcCellRef into multiple GcCellRefs for different components of the borrowed data.

The GcCell is already immutably borrowed, so this cannot fail.

This is an associated function that needs to be used as GcCellRef::map_split(...). A method would interfere with methods of the same name on the contents of a GcCellRef used through Deref.