Struct boa_gc::internals::ephemeron_box::EphemeronBox

pub(crate) struct EphemeronBox<K: Trace + 'static, V: Trace + 'static> {
    pub(crate) header: EphemeronBoxHeader,
    data: UnsafeCell<Option<Data<K, V>>>,
}

The inner allocation of an Ephemeron pointer.

Fields
header: EphemeronBoxHeaderdata: UnsafeCell<Option<Data<K, V>>>
Implementations

impl<K: Trace, V: Trace> EphemeronBox<K, V>

pub(crate) fn new(key: &Gc<K>, value: V) -> Self

pub(crate) fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two references refer to the same EphemeronBox.

pub(crate) unsafe fn value(&self) -> Option<&V>

Returns a reference to the ephemeron’s value or None.

Safety

The garbage collector must not run between the call to this function and the eventual drop of the returned reference, since that could free the inner value.

pub(crate) unsafe fn key(&self) -> Option<&GcBox<K>>

Returns a reference to the ephemeron’s key or None.

Safety

The garbage collector must not run between the call to this function and the eventual drop of the returned reference, since that could free the inner value.

pub(crate) unsafe fn mark(&self)

Marks this EphemeronBox as live.

This doesn’t mark the inner value of the ephemeron. ErasedEphemeronBox::trace does this, and it’s called by the garbage collector on demand.

pub(crate) fn inc_ref_count(&self)

pub(crate) fn dec_ref_count(&self)

pub(crate) fn inc_non_root_count(&self)