Struct boa_gc::internals::gc_box::GcBoxHeader

pub(crate) struct GcBoxHeader {
    roots: Cell<usize>,
    pub(crate) next: Cell<Option<NonNull<GcBox<dyn Trace>>>>,
}

The GcBoxheader contains the GcBox’s current state for the Collector’s Mark/Sweep as well as a pointer to the next node in the heap.

These flags include:

Root Count

Mark Flag Bit

Weak Flag Bit

The next node is set by the Allocator during initialization and by the Collector during the sweep phase.

Fields
roots: Cell<usize>next: Cell<Option<NonNull<GcBox<dyn Trace>>>>
Implementations

impl GcBoxHeader

pub(crate) fn new() -> Self

Creates a new GcBoxHeader with a root of 1 and next set to None.

pub(crate) fn new_weak() -> Self

Creates a new GcBoxHeader with the Weak bit at 1 and roots of 1.

pub(crate) fn roots(&self) -> usize

Returns the GcBoxHeader’s current root count

pub(crate) fn inc_roots(&self)

Increments GcBoxHeader’s root count.

pub(crate) fn dec_roots(&self)

Decreases GcBoxHeader’s current root count.

pub(crate) fn is_marked(&self) -> bool

Returns a bool for whether GcBoxHeader’s mark bit is 1.

pub(crate) fn mark(&self)

Sets GcBoxHeader’s mark bit to 1.

pub(crate) fn unmark(&self)

Sets GcBoxHeader’s mark bit to 0.

pub(crate) fn is_ephemeron(&self) -> bool

Returns a bool for whether the GcBoxHeader’s weak bit is 1.