Struct boa_gc::pointers::ephemeron::Ephemeron

pub struct Ephemeron<K: Trace + ?Sized + 'static, V: Trace + 'static> {
    inner_ptr: Cell<Rootable<EphemeronBox<K, V>>>,
}

A key-value pair where the value becomes unaccesible when the key is garbage collected.

See Racket’s explanation on ephemerons for a brief overview or read Barry Hayes’ Ephemerons: a new finalization mechanism.

Fields
inner_ptr: Cell<Rootable<EphemeronBox<K, V>>>
Implementations

impl<K: Trace + ?Sized, V: Trace + Clone> Ephemeron<K, V>

pub fn value(&self) -> Option<V>

Gets the stored value of this Ephemeron, or None if the key was already garbage collected.

This needs to return a clone of the value because holding a reference to it between garbage collection passes could drop the underlying allocation, causing an Use After Free.

pub fn has_value(&self) -> bool

Checks if the Ephemeron has a value.

impl<K: Trace + ?Sized, V: Trace> Ephemeron<K, V>

pub fn new(key: &Gc<K>, value: V) -> Self

Creates a new Ephemeron.

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two Ephemerons point to the same allocation.

fn is_rooted(&self) -> bool

fn root_ptr(&self)

fn unroot_ptr(&self)

pub(crate) fn inner_ptr(&self) -> NonNull<EphemeronBox<K, V>>

fn inner(&self) -> &EphemeronBox<K, V>

unsafe fn from_raw(ptr: NonNull<EphemeronBox<K, V>>) -> Self

Constructs an Ephemeron<K, V> from a raw pointer.

Safety

This function is unsafe because improper use may lead to memory corruption, double-free, or misbehaviour of the garbage collector.