Struct boa_gc::pointers::gc::Gc

pub struct Gc<T: Trace + ?Sized + 'static> {
    pub(crate) inner_ptr: NonNull<GcBox<T>>,
    pub(crate) marker: PhantomData<Rc<T>>,
}

A garbage-collected pointer type over an immutable value.

Fields
inner_ptr: NonNull<GcBox<T>>marker: PhantomData<Rc<T>>
Implementations

impl<T: Trace> Gc<T>

pub fn new(value: T) -> Self

Constructs a new Gc<T> with the given value.

pub fn new_cyclic<F>(data_fn: F) -> Selfwhere F: FnOnce(&WeakGc<T>) -> T,

Constructs a new Gc<T> while giving you a WeakGc<T> to the allocation, to allow constructing a T which holds a weak pointer to itself.

Since the new Gc<T> is not fully-constructed until Gc<T>::new_cyclic returns, calling upgrade on the weak reference inside the closure will fail and result in a None value.

pub fn into_raw(this: Self) -> NonNull<GcBox<T>>

Consumes the Gc, returning a wrapped raw pointer.

To avoid a memory leak, the pointer must be converted back to a Gc using Gc::from_raw.

impl<T: Trace + ?Sized> Gc<T>

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two Gcs point to the same allocation.

pub const unsafe fn from_raw(inner_ptr: NonNull<GcBox<T>>) -> Self

Constructs a Gc<T> from a raw pointer.

The raw pointer must have been returned by a previous call to Gc<U>::into_raw where U must have the same size and alignment as T.

Safety

This function is unsafe because improper use may lead to memory corruption, double-free, or misbehaviour of the garbage collector.

impl<T: Trace + ?Sized> Gc<T>

pub(crate) fn inner_ptr(&self) -> NonNull<GcBox<T>>

fn inner(&self) -> &GcBox<T>