Struct boa_gc::pointers::weak_map::RawWeakMap

pub(crate) struct RawWeakMap<K, V, S = DefaultHashBuilder>where
    K: Trace + 'static,
    V: Trace + 'static,{
    hash_builder: S,
    table: RawTable<Ephemeron<K, V>>,
}

A hash map where the bucket type is an Ephemeron<K, V>.

This data structure allows associating a Gc<K> with a value V that will be invalidated when the Gc<K> gets collected. In other words, all key entries on the map are weakly held.

Fields
hash_builder: Stable: RawTable<Ephemeron<K, V>>
Implementations

impl<K, V> RawWeakMap<K, V, DefaultHashBuilder>where K: Trace + 'static, V: Trace + 'static,

pub(crate) fn new() -> Self

Creates an empty RawWeakMap.

The map is initially created with a capacity of 0, so it will not allocate until it is first inserted into.

pub(crate) fn with_capacity(capacity: usize) -> Self

Creates an empty RawWeakMap with the specified capacity.

The map will be able to hold at least capacity elements without reallocating. If capacity is 0, the map will not allocate.

impl<K, V, S> RawWeakMap<K, V, S>where K: Trace + 'static, V: Trace + 'static,

pub(crate) const fn with_hasher(hash_builder: S) -> Self

Creates an empty RawWeakMap which will use the given hash builder to hash keys.

The map is initially created with a capacity of 0, so it will not allocate until it is first inserted into.

pub(crate) fn with_capacity_and_hasher(capacity: usize, hash_builder: S) -> Self

Creates an empty RawWeakMap with the specified capacity, using hash_builder to hash the keys.

The map will be able to hold at least capacity elements without reallocating. If capacity is 0, the map will not allocate.

pub(crate) const fn hasher(&self) -> &S

Returns a reference to the map’s BuildHasher.

pub(crate) fn capacity(&self) -> usize

Returns the number of elements the map can hold without reallocating.

This number is a lower bound; the map might be able to hold more, but is guaranteed to be able to hold at least this many.

pub(crate) fn iter(&self) -> Iter<'_, K, V>

An iterator visiting all entries in arbitrary order. The iterator element type is Ephemeron<K, V>.

pub(crate) fn len(&self) -> usize

Returns the number of elements in the map.

This is an upper bound; the map might contain some expired keys which haven’t been removed.

pub(crate) fn is_empty(&self) -> bool

Returns true if the map contains no elements.

This might return false if the map has expired keys that are still pending to be cleaned up.

pub(crate) fn retain<F>(&mut self, f: F)where F: FnMut(&Ephemeron<K, V>) -> bool,

Retains only the elements specified by the predicate. Keeps the allocated memory for reuse.

In other words, remove all ephemerons Ephemeron<K, V> such that f(&eph) returns false. The elements are visited in unsorted (and unspecified) order.

pub(crate) fn clear(&mut self)

Clears the map, removing all key-value pairs. Keeps the allocated memory for reuse.

impl<K, V, S> RawWeakMap<K, V, S>where K: Trace + 'static, V: Trace + Clone + 'static, S: BuildHasher,

pub(crate) fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the RawWeakMap. The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the new capacity exceeds isize::MAX bytes and abort the program in case of allocation error. Use try_reserve instead if you want to handle memory allocation failure.

pub(crate) fn try_reserve( &mut self, additional: usize ) -> Result<(), TryReserveError>

Tries to reserve capacity for at least additional more elements to be inserted in the given RawWeakMap<K,V>. The collection may reserve more space to avoid frequent reallocations.

Errors

If the capacity overflows, or the allocator reports a failure, then an error is returned.

pub(crate) fn shrink_to_fit(&mut self)

Shrinks the capacity of the map as much as possible. It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

pub(crate) fn shrink_to(&mut self, min_capacity: usize)

Shrinks the capacity of the map with a lower limit. It will drop down no lower than the supplied limit while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

This function does nothing if the current capacity is smaller than the supplied minimum capacity.

pub(crate) fn get(&self, k: &Gc<K>) -> Option<V>

Returns the value corresponding to the supplied key.

pub(crate) fn contains_key(&self, k: &Gc<K>) -> bool

Returns true if the map contains a value for the specified key.

pub(crate) fn insert(&mut self, k: &Gc<K>, v: V) -> Option<Ephemeron<K, V>>

If the map did not have this key present, None is returned.

If the map did have this key present, the value is updated, and the old value is returned. The key is not updated.

pub(crate) fn remove(&mut self, k: &Gc<K>) -> Option<V>

Removes a key from the map, returning the value at the key if the key was previously in the map. Keeps the allocated memory for reuse.

pub(crate) fn clear_expired(&mut self)

Clears all the expired keys in the map.