Struct boa_gc::Collector

pub(crate) struct Collector;

This collector currently functions in four main phases

Mark -> Finalize -> Mark -> Sweep

Mark nodes as reachable.

Finalize the unreachable nodes.

Mark again because Finalize::finalize can potentially resurrect dead nodes.

Sweep and drop all dead nodes.

A better approach in a more concurrent structure may be to reorder.

Mark -> Sweep -> Finalize

Implementations

impl Collector

pub(crate) fn collect(gc: &mut BoaGc)

Run a collection on the full heap.

pub(crate) fn mark_heap(
strong: &Cell<Option<NonNull<GcBox<dyn Trace>>>>,
weak: &Cell<Option<NonNull<dyn ErasedEphemeronBox>>>,
weak_map: &Cell<Option<NonNull<dyn ErasedWeakMapBox>>>
) -> Unreachables

Walk the heap and mark any nodes deemed reachable

pub(crate) unsafe fn finalize(unreachables: Unreachables)

Safety

Passing a strong or a weak vec with invalid pointers will result in Undefined Behaviour.

pub(crate) unsafe fn sweep(
strong: &Cell<Option<NonNull<GcBox<dyn Trace>>>>,
weak: &Cell<Option<NonNull<dyn ErasedEphemeronBox>>>,
total_allocated: &mut usize
)

Safety

Providing an invalid pointer in the heap_start or in any of the headers of each node will result in Undefined Behaviour.

Providing a list of pointers that weren’t allocated by Box::into_raw(Box::new(..)) will result in Undefined Behaviour.

pub(crate) fn dump(gc: &mut BoaGc)