Struct boa_gc::GcCell

pub struct GcCell<T: ?Sized + 'static> {
    pub(crate) flags: Cell<BorrowFlag>,
    pub(crate) cell: UnsafeCell<T>,
}

A mutable memory location with dynamically checked borrow rules that can be used inside of a garbage-collected pointer.

This object is a RefCell that can be used inside of a Gc<T>.

Fields
flags: Cell<BorrowFlag>cell: UnsafeCell<T>
Implementations

impl<T: Trace> GcCell<T>

pub const fn new(value: T) -> Self

Creates a new GcCell containing value.

pub fn into_inner(self) -> T

Consumes the GcCell, returning the wrapped value.

impl<T: Trace + ?Sized> GcCell<T>

pub fn borrow(&self) -> GcCellRef<'_, T>

Immutably borrows the wrapped value.

The borrow lasts until the returned GcCellRef exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the value is currently mutably borrowed.

pub fn borrow_mut(&self) -> GcCellRefMut<'_, T>

Mutably borrows the wrapped value.

The borrow lasts until the returned GcCellRefMut exits scope. The value cannot be borrowed while this borrow is active.

Panics

Panics if the value is currently borrowed.

pub fn try_borrow(&self) -> Result<GcCellRef<'_, T>, BorrowError>

Immutably borrows the wrapped value, returning an error if the value is currently mutably borrowed.

The borrow lasts until the returned GcCellRef exits scope. Multiple immutable borrows can be taken out at the same time.

This is the non-panicking variant of borrow.

Errors

Returns an Err if the value is currently mutably borrowed.

pub fn try_borrow_mut(&self) -> Result<GcCellRefMut<'_, T>, BorrowMutError>

Mutably borrows the wrapped value, returning an error if the value is currently borrowed.

The borrow lasts until the returned GcCellRefMut exits scope. The value cannot be borrowed while this borrow is active.

This is the non-panicking variant of borrow_mut.

Errors

Returns an Err if the value is currently borrowed.