Struct boa_gc::WeakGc

#[repr(transparent)]
pub struct WeakGc<T: Trace + ?Sized + 'static> {
    inner: Ephemeron<T, Gc<T>>,
}

A weak reference to a Gc.

This type allows keeping references to Gc managed values without keeping them alive for garbage collections. However, this also means WeakGc::upgrade could return None at any moment.

Fields
inner: Ephemeron<T, Gc<T>>
Implementations

impl<T: Trace> WeakGc<T>

pub fn new(value: &Gc<T>) -> Self

Creates a new weak pointer for a garbage collected value.

pub fn upgrade(&self) -> Option<Gc<T>>

Upgrade returns a Gc pointer for the internal value if valid, or None if the value was already garbage collected.