Module boa_gc::trace

Macros

fn_finalize_trace_group 🔒

fn_finalize_trace_one 🔒

simple_empty_finalize_trace 🔒

tuple_finalize_trace 🔒

type_arg_tuple_based_finalize_trace_impls 🔒

Traits

Finalize

Substitute for the Drop trait for garbage collected types.

Trace

The Trace trait, which needs to be implemented on garbage-collected objects.