Crate boa_gc

Boa’s boa_gc crate implements a garbage collector.

Crate Overview

boa_gc is a mark-sweep garbage collector that implements a Trace and Finalize trait for garbage collected values.

About Boa

Boa is an open-source, experimental ECMAScript Engine written in Rust for lexing, parsing and executing ECMAScript/JavaScript. Currently, Boa supports some of the language. More information can be viewed at Boa’s website.

Try out the most recent release with Boa’s live demo playground.

Boa Crates

boa_ast - Boa’s ECMAScript Abstract Syntax Tree.

boa_engine - Boa’s implementation of ECMAScript builtin objects and execution.

boa_gc - Boa’s garbage collector.

boa_interner - Boa’s string interner.

boa_parser - Boa’s lexer and parser.

boa_profiler - Boa’s code profiler.

boa_unicode - Boa’s Unicode identifier.

boa_icu_provider - Boa’s ICU4X data provider.

Modules

cell 🔒

A garbage collected cell implementation

internals 🔒

pointers 🔒

Pointers represents the External types returned by the Boa Garbage Collector

trace 🔒

Macros

custom_trace

Utility macro to manually implement Trace on a type.

empty_trace

Utility macro to define an empty implementation of Trace.

Structs

Allocator 🔒

The Allocator handles allocation of garbage collected values.

BoaGc 🔒

Collector 🔒

This collector currently functions in four main phases

DropGuard 🔒

DropGuard flags whether the Collector is currently running Collector::sweep() or Collector::dump()

Ephemeron

A key-value pair where the value becomes unaccesible when the key is garbage collected.

Gc

A garbage-collected pointer type over an immutable value.

GcBox

A garbage collected allocation.

GcConfig 🔒

GcRef

A wrapper type for an immutably borrowed value from a GcCell<T>.

GcRefCell

A mutable memory location with dynamically checked borrow rules that can be used inside of a garbage-collected pointer.

GcRefMut

A wrapper type for a mutably borrowed value from a GcCell<T>.

GcRuntimeData 🔒

Unreachables 🔒

WeakGc

A weak reference to a Gc.

WeakMap

A map that holds weak references to its keys and is traced by the garbage collector.

Constants

BOA_GC 🔒

GC_DROPPING 🔒

Statics

__KEY 🔒

__KEY 🔒

Traits

Finalize

Substitute for the Drop trait for garbage collected types.

Trace

The Trace trait, which needs to be implemented on garbage-collected objects.

Functions

__getit 🔒 ⚠

__getit 🔒 ⚠

__init 🔒

__init 🔒

finalizer_safe

Returns true if it is safe for a type to run Finalize::finalize.

force_collect

Forcefully runs a garbage collection of all unaccessible nodes.

Type Definitions

EphemeronPointer 🔒

ErasedWeakMapBoxPointer 🔒

GcPointer 🔒

Derive Macros

Finalize

Derive the Finalize trait.

Trace

Derive the Trace trait.