Crate boa_gc

Garbage collector for the Boa JavaScript engine.

Modules

cell 🔒

A garbage collected cell implementation

internals 🔒

pointers 🔒

Pointers represents the External types returned by the Boa Garbage Collector

trace 🔒

Macros

custom_trace

Utility macro to manually implement Trace on a type.

empty_trace

Utility macro to define an empty implementation of Trace.

Structs

Allocator 🔒

The Allocator handles allocation of garbage collected values.

BoaGc 🔒

Collector 🔒

DropGuard 🔒

DropGuard flags whether the Collector is currently running Collector::sweep() or Collector::dump()

Ephemeron

A key-value pair where the value becomes unaccesible when the key is garbage collected.

Gc

A garbage-collected pointer type over an immutable value.

GcCell

A mutable memory location with dynamically checked borrow rules that can be used inside of a garbage-collected pointer.

GcCellRef

A wrapper type for an immutably borrowed value from a GcCell<T>.

GcCellRefMut

A wrapper type for a mutably borrowed value from a GcCell<T>.

GcConfig 🔒

GcRuntimeData 🔒

WeakGc

A weak reference to a Gc.

Constants

BOA_GC 🔒

EPHEMERON_QUEUE 🔒

GC_DROPPING 🔒

Traits

Finalize

Substitute for the Drop trait for garbage collected types.

Trace

The Trace trait, which needs to be implemented on garbage-collected objects.

Functions

finalizer_safe

Returns true if it is safe for a type to run Finalize::finalize.

force_collect

Forcefully runs a garbage collection of all unaccessible nodes.

Type Definitions

GcPointer 🔒

Derive Macros

Finalize

Derive the Finalize trait.

Trace

Derive the Trace trait.