Crate boa_gc

Garbage collector for the Boa JavaScript engine.

Macros

custom_trace

This rule implements the trace method.

unsafe_empty_trace

This rule implements the trace methods with empty implementations.

Structs

Cell

A mutable memory location with dynamically checked borrow rules that can be used inside of a garbage-collected pointer.

Gc

A garbage-collected pointer type over an immutable value.

Ref

A wrapper type for an immutably borrowed value from a GcCell<T>.

RefMut

A wrapper type for a mutably borrowed value from a GcCell<T>.

Traits

Finalize

The Finalize trait, which needs to be implemented on garbage-collected objects to define finalization logic.

Trace

The Trace trait, which needs to be implemented on garbage-collected objects.

Functions

finalizer_safe

force_collect

Immediately triggers a garbage collection on the current thread.

Derive Macros

Finalize

Trace