Struct boa_gc::internals::ephemeron_box::EphemeronBox

pub(crate) struct EphemeronBox<K: Trace + ?Sized + 'static, V: Trace + ?Sized + 'static> {
    key: Cell<Option<NonNull<GcBox<K>>>>,
    value: V,
}

The inner allocation of an Ephemeron pointer.

Fields
key: Cell<Option<NonNull<GcBox<K>>>>value: V
Implementations

impl<K: Trace + ?Sized, V: Trace> EphemeronBox<K, V>

pub(crate) fn new(key: &Gc<K>, value: V) -> Self

impl<K: Trace + ?Sized, V: Trace + ?Sized> EphemeronBox<K, V>

pub(crate) fn is_marked(&self) -> bool

Checks if the key pointer is marked by Trace

pub(crate) fn inner_key_ptr(&self) -> Option<*mut GcBox<K>>

Returns some pointer to the key’s GcBox or None

Panics

This method will panic if called while the garbage collector is dropping.

pub(crate) fn inner_key(&self) -> Option<&GcBox<K>>

Returns some reference to key’s GcBox or None

pub(crate) fn key(&self) -> Option<&K>

Returns a reference to the value of key’s GcBox

pub(crate) const fn value(&self) -> &V

Returns a reference to value

fn weak_trace_key(&self)

Calls Trace::weak_trace() on key

fn weak_trace_value(&self)

Calls Trace::weak_trace() on value