Trait boa_gc::internals::ephemeron_box::ErasedEphemeronBox

pub(crate) trait ErasedEphemeronBox {
    // Required methods
    fn header(&self) -> &EphemeronBoxHeader;
    unsafe fn trace(&self) -> bool;
    fn trace_non_roots(&self);
    fn finalize_and_clear(&self);
}

Required Methods

fn header(&self) -> &EphemeronBoxHeader

Gets the header of the EphemeronBox.

unsafe fn trace(&self) -> bool

Traces through the EphemeronBox’s held value, but only if it’s marked and its key is also marked. Returns true if the ephemeron successfuly traced through its value. This also considers ephemerons that are marked but don’t have their value anymore as “successfully traced”.

fn trace_non_roots(&self)

fn finalize_and_clear(&self)

Runs the finalization logic of the EphemeronBox’s held value, if the key is still live, and clears its contents.