Struct boa_gc::pointers::gc::Gc

pub struct Gc<T: Trace + ?Sized + 'static> {
    pub(crate) inner_ptr: Cell<NonNull<GcBox<T>>>,
    pub(crate) marker: PhantomData<Rc<T>>,
}

A garbage-collected pointer type over an immutable value.

Fields
inner_ptr: Cell<NonNull<GcBox<T>>>marker: PhantomData<Rc<T>>
Implementations

impl<T: Trace> Gc<T>

pub fn new(value: T) -> Self

Constructs a new Gc<T> with the given value.

impl<T: Trace + ?Sized> Gc<T>

pub fn ptr_eq(this: &Self, other: &Self) -> bool

Returns true if the two Gcs point to the same allocation.

pub(crate) fn from_ptr(ptr: NonNull<GcBox<T>>) -> Self

Will return a new rooted Gc from a GcBox pointer

impl<T: Trace + ?Sized> Gc<T>

fn rooted(&self) -> bool

pub(crate) fn set_root(&self)

fn clear_root(&self)

pub(crate) fn inner_ptr(&self) -> NonNull<GcBox<T>>

fn inner(&self) -> &GcBox<T>